fastapi/app/requirements.txt
fastapi
pydantic
starlette
typing-extensions
uvicorn[standard]
gunicorn
requests
python-dotenv
pytz
python-multipart
SQLAlchemy
psycopg2
python-jose[cryptography]
passlib[bcrypt]
aiofiles
GeoAlchemy2==0.10.2
jinja2
fastapi-restful[tasks]
pandas
protobuf==3.19.6
google
shapely
pandas
# fastapi-cache2[redis]
# aioredis